Fixed map a bit

+++ src/main/java/com/maxminmajcdg/entities/CAVotesCong2018Entity.java
@@ -10,8 +10,8 @@ import javax.persistence.Table;
 @Table(name="cali_election18")
 public class CAVotesCong2018Entity extends VotesCongEntity { 
 	
-	public Map<ElectionCategory, Map<PartyCategory, Integer>> getPresVotes() {
-		Map<ElectionCategory, Map<PartyCategory, Integer>> election = new HashMap<ElectionCategory, Map<PartyCategory, Integer>>();
+	public Map<ElectionCategory, Map<PartyCategory, Double>> getPresVotes() {
+		Map<ElectionCategory, Map<PartyCategory, Double>> election = new HashMap<ElectionCategory, Map<PartyCategory, Double>>();
 		election.put(ElectionCategory.CONGRESSIONAL2016, getVotes());
 		return election;
 	}
